vxlan: Fix error on linux 6.5
With linux 6.5, the new `IFLA_VXLAN_LOCALBYPASS` will cause DecodeError
as current netlink-packet-route raise error on unknown Nla.

This patch will store the unknown Nla to `Other(DefaultNla)` to
compatibility.

Unit test case included.
