+++ src/rtnl/link/mod.rs
@@ -5,4 +5,6 @@ mod header;
 mod message;
 pub mod nlas;
 
+mod tests;
+
 pub use self::{buffer::*, header::*, message::*};
