+++ src/rtnl/link/nlas/link_infos.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 
 use anyhow::Context;
-use byteorder::{BigEndian, ByteOrder, NativeEndian};
+use byteorder::{ByteOrder, NativeEndian};
 use netlink_packet_utils::{
     nla::{DefaultNla, Nla, NlaBuffer, NlasIterator},
     parsers::{
@@ -13,7 +13,9 @@ use netlink_packet_utils::{
 };
 
 use super::{bond::InfoBond, bond_port::InfoBondPort, bridge::InfoBridge};
-use crate::{constants::*, LinkMessage, LinkMessageBuffer};
+use crate::{
+    constants::*, nlas::link::InfoVxlan, LinkMessage, LinkMessageBuffer,
+};
 
 const DUMMY: &str = "dummy";
 const IFB: &str = "ifb";
@@ -683,277 +685,6 @@ impl<'a, T: AsRef<[u8]> + ?Sized> Parseable<NlaBuffer<&'a T>> for InfoPortKind {
         Ok(match s.as_str() {
             BOND => Bond,
             _ => Other(s),
-        })
-    }
-}
-
-// https://elixir.bootlin.com/linux/v5.9.8/source/drivers/net/vxlan.c#L3332
-#[derive(Debug, PartialEq, Eq, Clone)]
-#[non_exhaustive]
-pub enum InfoVxlan {
-    Unspec(Vec<u8>),
-    Id(u32),
-    Group(Vec<u8>),
-    Group6(Vec<u8>),
-    Link(u32),
-    Local(Vec<u8>),
-    Local6(Vec<u8>),
-    Tos(u8),
-    Ttl(u8),
-    Label(u32),
-    Learning(u8),
-    Ageing(u32),
-    Limit(u32),
-    PortRange((u16, u16)),
-    Proxy(u8),
-    Rsc(u8),
-    L2Miss(u8),
-    L3Miss(u8),
-    CollectMetadata(u8),
-    Port(u16),
-    UDPCsum(u8),
-    UDPZeroCsumTX(u8),
-    UDPZeroCsumRX(u8),
-    RemCsumTX(u8),
-    RemCsumRX(u8),
-    Gbp(u8),
-    Gpe(u8),
-    RemCsumNoPartial(u8),
-    TtlInherit(u8),
-    Df(u8),
-}
-
-impl Nla for InfoVxlan {
-    #[rustfmt::skip]
-    fn value_len(&self) -> usize {
-        use self::InfoVxlan::*;
-        match *self {
-            Tos(_)
-                | Ttl(_)
-                | Learning(_)
-                | Proxy(_)
-                | Rsc(_)
-                | L2Miss(_)
-                | L3Miss(_)
-                | CollectMetadata(_)
-                | UDPCsum(_)
-                | UDPZeroCsumTX(_)
-                | UDPZeroCsumRX(_)
-                | RemCsumTX(_)
-                | RemCsumRX(_)
-                | Gbp(_)
-                | Gpe(_)
-                | RemCsumNoPartial(_)
-                | TtlInherit(_)
-                | Df(_)
-            => 1,
-            Port(_) => 2,
-            Id(_)
-                | Label(_)
-                | Link(_)
-                | Ageing(_)
-                | Limit(_)
-                | PortRange(_)
-            => 4,
-            Local(ref bytes)
-                | Local6(ref bytes)
-                | Group(ref bytes)
-                | Group6(ref bytes)
-                | Unspec(ref bytes)
-            => bytes.len(),
-        }
-    }
-
-    #[rustfmt::skip]
-    fn emit_value(&self, buffer: &mut [u8]) {
-        use self::InfoVxlan::*;
-        match self {
-            Unspec(ref bytes) => buffer.copy_from_slice(bytes),
-            Id(ref value)
-                | Label(ref value)
-                | Link(ref value)
-                | Ageing(ref value)
-                | Limit(ref value)
-            => NativeEndian::write_u32(buffer, *value),
-            Tos(ref value)
-                | Ttl(ref value)
-                | Learning (ref value)
-                | Proxy(ref value)
-                | Rsc(ref value)
-                | L2Miss(ref value)
-                | L3Miss(ref value)
-                | CollectMetadata(ref value)
-                | UDPCsum(ref value)
-                | UDPZeroCsumTX(ref value)
-                | UDPZeroCsumRX(ref value)
-                | RemCsumTX(ref value)
-                | RemCsumRX(ref value)
-                | Gbp(ref value)
-                | Gpe(ref value)
-                | RemCsumNoPartial(ref value)
-                | TtlInherit(ref value)
-                | Df(ref value)
-            =>  buffer[0] = *value,
-            Local(ref value)
-                | Group(ref value)
-                | Group6(ref value)
-                | Local6(ref value)
-            => buffer.copy_from_slice(value.as_slice()),
-            Port(ref value) => BigEndian::write_u16(buffer, *value),
-            PortRange(ref range) => {
-                BigEndian::write_u16(buffer, range.0);
-                BigEndian::write_u16(&mut buffer[2..], range.1)
-            }
-        }
-    }
-
-    fn kind(&self) -> u16 {
-        use self::InfoVxlan::*;
-
-        match self {
-            Id(_) => IFLA_VXLAN_ID,
-            Group(_) => IFLA_VXLAN_GROUP,
-            Group6(_) => IFLA_VXLAN_GROUP6,
-            Link(_) => IFLA_VXLAN_LINK,
-            Local(_) => IFLA_VXLAN_LOCAL,
-            Local6(_) => IFLA_VXLAN_LOCAL6,
-            Tos(_) => IFLA_VXLAN_TOS,
-            Ttl(_) => IFLA_VXLAN_TTL,
-            Label(_) => IFLA_VXLAN_LABEL,
-            Learning(_) => IFLA_VXLAN_LEARNING,
-            Ageing(_) => IFLA_VXLAN_AGEING,
-            Limit(_) => IFLA_VXLAN_LIMIT,
-            PortRange(_) => IFLA_VXLAN_PORT_RANGE,
-            Proxy(_) => IFLA_VXLAN_PROXY,
-            Rsc(_) => IFLA_VXLAN_RSC,
-            L2Miss(_) => IFLA_VXLAN_L2MISS,
-            L3Miss(_) => IFLA_VXLAN_L3MISS,
-            CollectMetadata(_) => IFLA_VXLAN_COLLECT_METADATA,
-            Port(_) => IFLA_VXLAN_PORT,
-            UDPCsum(_) => IFLA_VXLAN_UDP_CSUM,
-            UDPZeroCsumTX(_) => IFLA_VXLAN_UDP_ZERO_CSUM6_TX,
-            UDPZeroCsumRX(_) => IFLA_VXLAN_UDP_ZERO_CSUM6_RX,
-            RemCsumTX(_) => IFLA_VXLAN_REMCSUM_TX,
-            RemCsumRX(_) => IFLA_VXLAN_REMCSUM_RX,
-            Gbp(_) => IFLA_VXLAN_GBP,
-            Gpe(_) => IFLA_VXLAN_GPE,
-            RemCsumNoPartial(_) => IFLA_VXLAN_REMCSUM_NOPARTIAL,
-            TtlInherit(_) => IFLA_VXLAN_TTL_INHERIT,
-            Df(_) => IFLA_VXLAN_DF,
-            Unspec(_) => IFLA_VXLAN_UNSPEC,
-        }
-    }
-}
-
-impl<'a, T: AsRef<[u8]> + ?Sized> Parseable<NlaBuffer<&'a T>> for InfoVxlan {
-    fn parse(buf: &NlaBuffer<&'a T>) -> Result<Self, DecodeError> {
-        use self::InfoVxlan::*;
-        let payload = buf.value();
-        Ok(match buf.kind() {
-            IFLA_VLAN_UNSPEC => Unspec(payload.to_vec()),
-            IFLA_VXLAN_ID => {
-                Id(parse_u32(payload).context("invalid IFLA_VXLAN_ID value")?)
-            }
-            IFLA_VXLAN_GROUP => Group(payload.to_vec()),
-            IFLA_VXLAN_GROUP6 => Group6(payload.to_vec()),
-            IFLA_VXLAN_LINK => Link(
-                parse_u32(payload).context("invalid IFLA_VXLAN_LINK value")?,
-            ),
-            IFLA_VXLAN_LOCAL => Local(payload.to_vec()),
-            IFLA_VXLAN_LOCAL6 => Local6(payload.to_vec()),
-            IFLA_VXLAN_TOS => {
-                Tos(parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_TOS value")?)
-            }
-            IFLA_VXLAN_TTL => {
-                Ttl(parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_TTL value")?)
-            }
-            IFLA_VXLAN_LABEL => Label(
-                parse_u32(payload).context("invalid IFLA_VXLAN_LABEL value")?,
-            ),
-            IFLA_VXLAN_LEARNING => Learning(
-                parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_LEARNING value")?,
-            ),
-            IFLA_VXLAN_AGEING => Ageing(
-                parse_u32(payload)
-                    .context("invalid IFLA_VXLAN_AGEING value")?,
-            ),
-            IFLA_VXLAN_LIMIT => Limit(
-                parse_u32(payload).context("invalid IFLA_VXLAN_LIMIT value")?,
-            ),
-            IFLA_VXLAN_PROXY => Proxy(
-                parse_u8(payload).context("invalid IFLA_VXLAN_PROXY value")?,
-            ),
-            IFLA_VXLAN_RSC => {
-                Rsc(parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_RSC value")?)
-            }
-            IFLA_VXLAN_L2MISS => L2Miss(
-                parse_u8(payload).context("invalid IFLA_VXLAN_L2MISS value")?,
-            ),
-            IFLA_VXLAN_L3MISS => L3Miss(
-                parse_u8(payload).context("invalid IFLA_VXLAN_L3MISS value")?,
-            ),
-            IFLA_VXLAN_COLLECT_METADATA => CollectMetadata(
-                parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_COLLECT_METADATA value")?,
-            ),
-            IFLA_VXLAN_PORT_RANGE => {
-                let err = "invalid IFLA_VXLAN_PORT value";
-                if payload.len() != 4 {
-                    return Err(err.into());
-                }
-                let low = parse_u16_be(&payload[0..2]).context(err)?;
-                let high = parse_u16_be(&payload[2..]).context(err)?;
-                PortRange((low, high))
-            }
-            IFLA_VXLAN_PORT => Port(
-                parse_u16_be(payload)
-                    .context("invalid IFLA_VXLAN_PORT value")?,
-            ),
-            IFLA_VXLAN_UDP_CSUM => UDPCsum(
-                parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_UDP_CSUM value")?,
-            ),
-            IFLA_VXLAN_UDP_ZERO_CSUM6_TX => UDPZeroCsumTX(
-                parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_UDP_ZERO_CSUM6_TX value")?,
-            ),
-            IFLA_VXLAN_UDP_ZERO_CSUM6_RX => UDPZeroCsumRX(
-                parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_UDP_ZERO_CSUM6_RX value")?,
-            ),
-            IFLA_VXLAN_REMCSUM_TX => RemCsumTX(
-                parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_REMCSUM_TX value")?,
-            ),
-            IFLA_VXLAN_REMCSUM_RX => RemCsumRX(
-                parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_REMCSUM_RX value")?,
-            ),
-            IFLA_VXLAN_DF => {
-                Df(parse_u8(payload).context("invalid IFLA_VXLAN_DF value")?)
-            }
-            IFLA_VXLAN_GBP => {
-                Gbp(parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_GBP value")?)
-            }
-            IFLA_VXLAN_GPE => {
-                Gpe(parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_GPE value")?)
-            }
-            IFLA_VXLAN_REMCSUM_NOPARTIAL => RemCsumNoPartial(
-                parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_REMCSUM_NO_PARTIAL")?,
-            ),
-            IFLA_VXLAN_TTL_INHERIT => TtlInherit(
-                parse_u8(payload)
-                    .context("invalid IFLA_VXLAN_TTL_INHERIT value")?,
-            ),
-            __IFLA_VXLAN_MAX => Unspec(payload.to_vec()),
-            _ => return Err(format!("unknown NLA type {}", buf.kind()).into()),
         })
     }
 }
