+++ src/rtnl/link/tests/mod.rs
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: MIT
+
+#[cfg(test)]
+mod vxlan;
